+++ perllib/FixMyStreet/Cobrand/Peterborough.pm
@@ -12,6 +12,7 @@ use Utils;
 use Moo;
 with 'FixMyStreet::Roles::ConfirmOpen311';
 with 'FixMyStreet::Roles::ConfirmValidation';
+with 'FixMyStreet::Roles::Open311Multi';
 
 sub council_area_id { 2566 }
 sub council_area { 'Peterborough' }
@@ -104,9 +105,6 @@ sub open311_munge_update_params {
 
     # Send the FMS problem ID with the update.
     $params->{service_request_id_ext} = $comment->problem->id;
-
-    my $contact = $comment->problem->contact;
-    $params->{service_code} = $contact->email;
 }
 
 around 'open311_config' => sub {
@@ -662,9 +660,16 @@ sub bin_services_for_address {
             report_open => ( @report_service_ids_open || $open_requests->{492} ) ? 1 : 0,
         };
         if ($row->{report_allowed}) {
+            # We only get here if we're within the 2.5 day window after the collection.
+            # Set this so missed food collections can always be reported, as they don't
+            # have their own collection event.
+            $self->{c}->stash->{any_report_allowed} = 1;
+
             # If on the day, but before 5pm, show a special message to call
+            # (which is slightly different for staff, who are actually allowed to report)
             if ($last->ymd eq $now->ymd && $now->hour < 17) {
-                $row->{report_allowed} = 0;
+                my $is_staff = $self->{c}->user_exists && $self->{c}->user->from_body && $self->{c}->user->from_body->name eq "Peterborough City Council";
+                $row->{report_allowed} = $is_staff ? 1 : 0;
                 $row->{report_locked_out} = "ON DAY PRE 5PM";
             }
             # But if it has been marked as locked out, show that
@@ -696,6 +701,7 @@ sub bin_services_for_address {
         request_only => 1,
         report_only => 1,
     };
+
     # We want this one to always appear first
     unshift @out, {
         id => "_ALL_BINS",
@@ -810,6 +816,16 @@ sub waste_munge_report_form_data {
     }
 }
 
+sub waste_munge_report_form_fields {
+    my ($self, $field_list) = @_;
+
+    push @$field_list, "extra_detail" => {
+        type => 'Text',
+        widget => 'Textarea',
+        label => 'Please supply additional information such as the location of the bin.',
+    };
+}
+
 sub waste_munge_request_data {
     my ($self, $id, $data) = @_;
 
@@ -865,6 +881,10 @@ sub waste_munge_report_data {
         $data->{detail} = $c->stash->{property}->{address};
     }
 
+    if ( $data->{extra_detail} ) {
+        $data->{detail} .= "\n\nExtra detail: " . $data->{extra_detail};
+    }
+
     $data->{category} = $self->body->contacts->find({ email => "Bartec-$service_id" })->category;
 }
 
@@ -893,6 +913,10 @@ sub waste_munge_enquiry_data {
     $data->{category} = $category;
     $data->{title} = $bin;
     $data->{detail} = $category_verbose . "\n\n" . $c->stash->{property}->{address};
+
+    if ( $data->{extra_extra_detail} ) {
+        $data->{detail} .= "\n\nExtra detail: " . $data->{extra_extra_detail};
+    }
 }