+++ perllib/Open311.pm
@@ -416,7 +416,8 @@ sub _populate_service_request_update_params {
 
     my $service_request_id = $comment->problem->external_id;
     if ( $self->use_customer_reference ) {
-        $service_request_id = $comment->problem->get_extra_metadata('customer_reference');
+        my $customer_ref = $comment->problem->get_extra_metadata('customer_reference');
+        $service_request_id = $customer_ref || $service_request_id;
     }
     my $params = {
         updated_datetime => DateTime::Format::W3CDTF->format_datetime($comment->confirmed->set_nanosecond(0)),
@@ -485,7 +486,6 @@ sub _request {
     my $path = shift;
     my $params = shift || {};
     my $uploads = shift;
-
     my $uri = URI->new( $self->endpoint );
     $uri->path( $uri->path . $path );