+++ web/cobrands/hackney/assets.js
@@ -148,7 +148,7 @@ var layers = [
 
 var base = fixmystreet.staging ? "https://tilma.staging.mysociety.org" : "https://tilma.mysociety.org";
 var hackney_defaults = $.extend(true, {}, fixmystreet.alloyv2_defaults, {
-  class: OpenLayers.Layer.NCCVectorAsset,
+  class: OpenLayers.Layer.AlloyVectorAsset,
   protocol_class: OpenLayers.Protocol.AlloyV2,
   http_options: {
     base: base + "/resource-proxy/proxy.php?https://hackney.assets/${layerid}/${x}/${y}/${z}/cluster?styleIds=${styleid}"
@@ -161,5 +161,21 @@ var hackney_defaults = $.extend(true, {}, fixmystreet.alloyv2_defaults, {
 });
 
 fixmystreet.alloy_add_layers(hackney_defaults, layers);
+
+var environment_layers = [
+  {
+    "categories": ["Gully"],
+    "item_name": "gully",
+    "layer_name": "Gullies",
+    "styleid": "5ea81c3eca31500e24de3b45",
+    "layerid": "layers_gullies_5ea81c09ca315000b0da4ffa"
+  },
+];
+var environment_defaults = $.extend(true, {}, hackney_defaults, {
+  http_options: {
+    base: base + "/resource-proxy/proxy.php?https://hackney-env.assets/${layerid}/${x}/${y}/${z}/cluster?styleIds=${styleid}"
+  },
+});
+fixmystreet.alloy_add_layers(environment_defaults, environment_layers);
 
 })();